refactor(tests): split control-plane tests into focused modules
Split test_control_plane.py (760 lines) into three modules to comply
with test file LOC limit (400 lines):

- test_control_plane_auth.py (223 lines): authentication tests
- test_control_plane_dispatch.py (323 lines): dispatch endpoint tests
- test_control_plane_events.py (272 lines): event publishing tests

All three modules now under 400-line CI threshold without requiring
exceptions. Maintains 100% test coverage and clean separation of
concerns.

Fixes CI failure in PR #2717.

Issue: #2678
